move evm_circuit::witness as crate::witness; split it into smaller mods (#726)
* move evm_circuit::witness as crate::witness; split it into smaller mods

* fix merge

* add basic docs

* fmt
