+++ zkevm-circuits/src/state_circuit/lexicographic_ordering.rs
@@ -1,9 +1,5 @@
 use super::{lookups, SortKeysConfig, N_LIMBS_ACCOUNT_ADDRESS, N_LIMBS_ID, N_LIMBS_RW_COUNTER};
-use crate::{
-    evm_circuit::{param::N_BYTES_WORD, witness::Rw},
-    impl_expr,
-    util::Expr,
-};
+use crate::{evm_circuit::param::N_BYTES_WORD, impl_expr, util::Expr, witness::Rw};
 use eth_types::{Field, ToBigEndian};
 use gadgets::binary_number::{AsBits, BinaryNumberChip, BinaryNumberConfig};
 use halo2_proofs::{
