+++ zkevm-circuits/src/witness/step.rs
@@ -1,0 +1,240 @@
+use bus_mapping::{
+    circuit_input_builder,
+    error::{ExecError, OogError},
+    evm::OpcodeId,
+    operation,
+};
+
+use crate::{
+    evm_circuit::{
+        param::{N_BYTES_WORD, STACK_CAPACITY},
+        step::ExecutionState,
+    },
+    table::RwTableTag,
+};
+
+/// Step executed in a transaction
+#[derive(Clone, Debug, Default, PartialEq, Eq)]
+pub struct ExecStep {
+    /// The index in the Transaction calls
+    pub call_index: usize,
+    /// The indices in the RW trace incurred in this step
+    pub rw_indices: Vec<(RwTableTag, usize)>,
+    /// The execution state for the step
+    pub execution_state: ExecutionState,
+    /// The Read/Write counter before the step
+    pub rw_counter: usize,
+    /// The program counter
+    pub program_counter: u64,
+    /// The stack pointer
+    pub stack_pointer: usize,
+    /// The amount of gas left
+    pub gas_left: u64,
+    /// The gas cost in this step
+    pub gas_cost: u64,
+    /// The memory size in bytes
+    pub memory_size: u64,
+    /// The counter for reversible writes
+    pub reversible_write_counter: usize,
+    /// The counter for log index within tx
+    pub log_id: usize,
+    /// The opcode corresponds to the step
+    pub opcode: Option<OpcodeId>,
+}
+
+impl ExecStep {
+    /// The memory size in word **before** this step
+    pub fn memory_word_size(&self) -> u64 {
+        // EVM always pads the memory size to word size
+        // https://github.com/ethereum/go-ethereum/blob/master/core/vm/interpreter.go#L212-L216
+        // Thus, the memory size must be a multiple of 32 bytes.
+        assert_eq!(self.memory_size % N_BYTES_WORD as u64, 0);
+        self.memory_size / N_BYTES_WORD as u64
+    }
+}
+
+impl From<&ExecError> for ExecutionState {
+    fn from(error: &ExecError) -> Self {
+        match error {
+            ExecError::InvalidOpcode => ExecutionState::ErrorInvalidOpcode,
+            ExecError::StackOverflow => ExecutionState::ErrorStackOverflow,
+            ExecError::StackUnderflow => ExecutionState::ErrorStackUnderflow,
+            ExecError::WriteProtection => ExecutionState::ErrorWriteProtection,
+            ExecError::Depth => ExecutionState::ErrorDepth,
+            ExecError::InsufficientBalance => ExecutionState::ErrorInsufficientBalance,
+            ExecError::ContractAddressCollision => ExecutionState::ErrorContractAddressCollision,
+            ExecError::InvalidCreationCode => ExecutionState::ErrorInvalidCreationCode,
+            ExecError::InvalidJump => ExecutionState::ErrorInvalidJump,
+            ExecError::ReturnDataOutOfBounds => ExecutionState::ErrorReturnDataOutOfBound,
+            ExecError::CodeStoreOutOfGas => ExecutionState::ErrorOutOfGasCodeStore,
+            ExecError::MaxCodeSizeExceeded => ExecutionState::ErrorMaxCodeSizeExceeded,
+            ExecError::OutOfGas(oog_error) => match oog_error {
+                OogError::Constant => ExecutionState::ErrorOutOfGasConstant,
+                OogError::StaticMemoryExpansion => {
+                    ExecutionState::ErrorOutOfGasStaticMemoryExpansion
+                }
+                OogError::DynamicMemoryExpansion => {
+                    ExecutionState::ErrorOutOfGasDynamicMemoryExpansion
+                }
+                OogError::MemoryCopy => ExecutionState::ErrorOutOfGasMemoryCopy,
+                OogError::AccountAccess => ExecutionState::ErrorOutOfGasAccountAccess,
+                OogError::CodeStore => ExecutionState::ErrorOutOfGasCodeStore,
+                OogError::Log => ExecutionState::ErrorOutOfGasLOG,
+                OogError::Exp => ExecutionState::ErrorOutOfGasEXP,
+                OogError::Sha3 => ExecutionState::ErrorOutOfGasSHA3,
+                OogError::ExtCodeCopy => ExecutionState::ErrorOutOfGasEXTCODECOPY,
+                OogError::Sload => ExecutionState::ErrorOutOfGasSLOAD,
+                OogError::Sstore => ExecutionState::ErrorOutOfGasSSTORE,
+                OogError::Call => ExecutionState::ErrorOutOfGasCALL,
+                OogError::CallCode => ExecutionState::ErrorOutOfGasCALLCODE,
+                OogError::DelegateCall => ExecutionState::ErrorOutOfGasDELEGATECALL,
+                OogError::Create2 => ExecutionState::ErrorOutOfGasCREATE2,
+                OogError::StaticCall => ExecutionState::ErrorOutOfGasSTATICCALL,
+                OogError::SelfDestruct => ExecutionState::ErrorOutOfGasSELFDESTRUCT,
+            },
+        }
+    }
+}
+
+impl From<&circuit_input_builder::ExecStep> for ExecutionState {
+    fn from(step: &circuit_input_builder::ExecStep) -> Self {
+        if let Some(error) = step.error.as_ref() {
+            return error.into();
+        }
+        match step.exec_state {
+            circuit_input_builder::ExecState::Op(op) => {
+                if op.is_dup() {
+                    return ExecutionState::DUP;
+                }
+                if op.is_push() {
+                    return ExecutionState::PUSH;
+                }
+                if op.is_swap() {
+                    return ExecutionState::SWAP;
+                }
+                if op.is_log() {
+                    return ExecutionState::LOG;
+                }
+
+                macro_rules! dummy {
+                    ($name:expr) => {{
+                        log::warn!("{:?} is implemented with DummyGadget", $name);
+                        $name
+                    }};
+                }
+
+                match op {
+                    OpcodeId::ADD | OpcodeId::SUB => ExecutionState::ADD_SUB,
+                    OpcodeId::ADDMOD => ExecutionState::ADDMOD,
+                    OpcodeId::ADDRESS => ExecutionState::ADDRESS,
+                    OpcodeId::MUL | OpcodeId::DIV | OpcodeId::MOD => ExecutionState::MUL_DIV_MOD,
+                    OpcodeId::MULMOD => ExecutionState::MULMOD,
+                    OpcodeId::SDIV | OpcodeId::SMOD => ExecutionState::SDIV_SMOD,
+                    OpcodeId::EQ | OpcodeId::LT | OpcodeId::GT => ExecutionState::CMP,
+                    OpcodeId::SLT | OpcodeId::SGT => ExecutionState::SCMP,
+                    OpcodeId::SIGNEXTEND => ExecutionState::SIGNEXTEND,
+                    OpcodeId::STOP => ExecutionState::STOP,
+                    OpcodeId::AND => ExecutionState::BITWISE,
+                    OpcodeId::XOR => ExecutionState::BITWISE,
+                    OpcodeId::OR => ExecutionState::BITWISE,
+                    OpcodeId::NOT => ExecutionState::NOT,
+                    OpcodeId::POP => ExecutionState::POP,
+                    OpcodeId::PUSH32 => ExecutionState::PUSH,
+                    OpcodeId::BYTE => ExecutionState::BYTE,
+                    OpcodeId::MLOAD => ExecutionState::MEMORY,
+                    OpcodeId::MSTORE => ExecutionState::MEMORY,
+                    OpcodeId::MSTORE8 => ExecutionState::MEMORY,
+                    OpcodeId::JUMPDEST => ExecutionState::JUMPDEST,
+                    OpcodeId::JUMP => ExecutionState::JUMP,
+                    OpcodeId::JUMPI => ExecutionState::JUMPI,
+                    OpcodeId::GASPRICE => ExecutionState::GASPRICE,
+                    OpcodeId::PC => ExecutionState::PC,
+                    OpcodeId::MSIZE => ExecutionState::MSIZE,
+                    OpcodeId::CALLER => ExecutionState::CALLER,
+                    OpcodeId::CALLVALUE => ExecutionState::CALLVALUE,
+                    OpcodeId::EXTCODEHASH => ExecutionState::EXTCODEHASH,
+                    OpcodeId::BLOCKHASH => ExecutionState::BLOCKHASH,
+                    OpcodeId::TIMESTAMP | OpcodeId::NUMBER | OpcodeId::GASLIMIT => {
+                        ExecutionState::BLOCKCTXU64
+                    }
+                    OpcodeId::COINBASE => ExecutionState::BLOCKCTXU160,
+                    OpcodeId::DIFFICULTY | OpcodeId::BASEFEE => ExecutionState::BLOCKCTXU256,
+                    OpcodeId::GAS => ExecutionState::GAS,
+                    OpcodeId::SELFBALANCE => ExecutionState::SELFBALANCE,
+                    OpcodeId::SHA3 => ExecutionState::SHA3,
+                    OpcodeId::SHL | OpcodeId::SHR => ExecutionState::SHL_SHR,
+                    OpcodeId::SLOAD => ExecutionState::SLOAD,
+                    OpcodeId::SSTORE => ExecutionState::SSTORE,
+                    OpcodeId::CALLDATASIZE => ExecutionState::CALLDATASIZE,
+                    OpcodeId::CALLDATACOPY => ExecutionState::CALLDATACOPY,
+                    OpcodeId::CHAINID => ExecutionState::CHAINID,
+                    OpcodeId::ISZERO => ExecutionState::ISZERO,
+                    OpcodeId::CALL => ExecutionState::CALL,
+                    OpcodeId::ORIGIN => ExecutionState::ORIGIN,
+                    OpcodeId::CODECOPY => ExecutionState::CODECOPY,
+                    OpcodeId::CALLDATALOAD => ExecutionState::CALLDATALOAD,
+                    OpcodeId::CODESIZE => ExecutionState::CODESIZE,
+                    OpcodeId::RETURN | OpcodeId::REVERT => ExecutionState::RETURN,
+                    // dummy ops
+                    OpcodeId::BALANCE => dummy!(ExecutionState::BALANCE),
+                    OpcodeId::EXP => dummy!(ExecutionState::EXP),
+                    OpcodeId::SAR => dummy!(ExecutionState::SAR),
+                    OpcodeId::EXTCODESIZE => dummy!(ExecutionState::EXTCODESIZE),
+                    OpcodeId::EXTCODECOPY => dummy!(ExecutionState::EXTCODECOPY),
+                    OpcodeId::RETURNDATASIZE => dummy!(ExecutionState::RETURNDATASIZE),
+                    OpcodeId::RETURNDATACOPY => dummy!(ExecutionState::RETURNDATACOPY),
+                    OpcodeId::CREATE => dummy!(ExecutionState::CREATE),
+                    OpcodeId::CALLCODE => dummy!(ExecutionState::CALLCODE),
+                    OpcodeId::DELEGATECALL => dummy!(ExecutionState::DELEGATECALL),
+                    OpcodeId::CREATE2 => dummy!(ExecutionState::CREATE2),
+                    OpcodeId::STATICCALL => dummy!(ExecutionState::STATICCALL),
+                    OpcodeId::SELFDESTRUCT => dummy!(ExecutionState::SELFDESTRUCT),
+                    _ => unimplemented!("unimplemented opcode {:?}", op),
+                }
+            }
+            circuit_input_builder::ExecState::BeginTx => ExecutionState::BeginTx,
+            circuit_input_builder::ExecState::EndTx => ExecutionState::EndTx,
+        }
+    }
+}
+
+pub(super) fn step_convert(step: &circuit_input_builder::ExecStep) -> ExecStep {
+    ExecStep {
+        call_index: step.call_index,
+        rw_indices: step
+            .bus_mapping_instance
+            .iter()
+            .map(|x| {
+                let tag = match x.target() {
+                    operation::Target::Memory => RwTableTag::Memory,
+                    operation::Target::Stack => RwTableTag::Stack,
+                    operation::Target::Storage => RwTableTag::AccountStorage,
+                    operation::Target::TxAccessListAccount => RwTableTag::TxAccessListAccount,
+                    operation::Target::TxAccessListAccountStorage => {
+                        RwTableTag::TxAccessListAccountStorage
+                    }
+                    operation::Target::TxRefund => RwTableTag::TxRefund,
+                    operation::Target::Account => RwTableTag::Account,
+                    operation::Target::AccountDestructed => RwTableTag::AccountDestructed,
+                    operation::Target::CallContext => RwTableTag::CallContext,
+                    operation::Target::TxReceipt => RwTableTag::TxReceipt,
+                    operation::Target::TxLog => RwTableTag::TxLog,
+                };
+                (tag, x.as_usize())
+            })
+            .collect(),
+        execution_state: ExecutionState::from(step),
+        rw_counter: usize::from(step.rwc),
+        program_counter: usize::from(step.pc) as u64,
+        stack_pointer: STACK_CAPACITY - step.stack_size,
+        gas_left: step.gas_left.0,
+        gas_cost: step.gas_cost.as_u64(),
+        opcode: match step.exec_state {
+            circuit_input_builder::ExecState::Op(op) => Some(op),
+            _ => None,
+        },
+        memory_size: step.memory_size as u64,
+        reversible_write_counter: step.reversible_write_counter,
+        log_id: step.log_id,
+    }
+}
